animalLookup , animalLookupDetail 소스파일 통합
