+++ AnimalHospitalProject/src/main/java/kr/human/anihospital/controller/AnimallookupController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import kr.human.anihospital.service.AnimallookupService;
+import kr.human.anihospital.vo.AnimalLookupDetailVO;
 import kr.human.anihospital.vo.AnimallookupVO;
 import lombok.extern.slf4j.Slf4j;
 
@@ -24,7 +25,7 @@ public class AnimallookupController {
 	// 한 명의 보호자에 따른 환자의 진료내역 리스트를 화면에 표시해줄 메서드
 	//----------------------------------------------------------------------------------------------------
 	@GetMapping(value = "/animallookup")
-	public String selectAnimallookup(@RequestParam(required = false, defaultValue = "11") int seqProtector, Model model) throws Exception{
+	public String selectAnimallookup(@RequestParam(required = false, defaultValue = "2") int seqProtector, Model model) throws Exception{
 		// animallookup에 띄울 값 표시 시작 ---------------------------------------------------------------------
 		
 		// 화면에 넘길 데이터를 담을 그릇 준비
@@ -39,5 +40,30 @@ public class AnimallookupController {
 		// animallookup에 띄울 값 표시 종료 ---------------------------------------------------------------------
 		
 		return "animallookup";
+	}
+	
+	//----------------------------------------------------------------------------------------------------
+	// 한 명의 보호자에 따른 특정 환자의 진료내역 리스트를 화면에 표시해줄 메서드
+	//----------------------------------------------------------------------------------------------------
+	@GetMapping(value = "/animallookupdetail")
+	public String selectAnimalLookupDetail(@RequestParam int seqAnimal, @RequestParam int seqProtector, Model model) throws Exception{
+		// animallookupdetail에 띄울 값 표시 시작 ---------------------------------------------------------------------
+		
+		// 제대로 데이터가 담겨 있는지 로그에 찍어보기
+		log.info("selectAnimalLookupDetail 컨트롤러 받은 값 seq : {}, {}", seqAnimal, seqProtector);
+		
+		// 화면에 넘길 데이터를 담을 그릇 준비
+		List<AnimalLookupDetailVO> animalLookupDetail = null;
+		// 데이터 그릇에 담기
+		animalLookupDetail = animallookupService.selectAnimalLookupDetail(seqAnimal, seqProtector);
+		// 받아온 데이터 화면에 넘겨주기
+		model.addAttribute("animalLookupDetail", animalLookupDetail);
+		
+		// 제대로 데이터가 담겨 있는지 로그에 찍어보기
+		log.info("selectAnimalLookupDetail 서비스에서 넘어온 값(컨트롤러) : {}", animalLookupDetail);
+		
+		// animallookupdetail에 띄울 값 표시 종료 ---------------------------------------------------------------------
+		
+		return "animallookupdetail";
 	}
 }
